Release retro-g15 emulator version 0.03.
. Correct determination of the return address in MARK EXIT. It was
working for immediate commands but not deferred ones.
. Correct detection of overflow for single- and double-precision
addition/subtraction.
. Correct inhibition of -0 result from single- and double-precision
addition/subtraction, as well as its interaction with overflow
detection.
. Do not allow -0 to reach the adder in single- and double-precision
addition -- the minus sign messes up the G-15's sign determination.
. Correct handling of PN sign for double-precision addition/subtraction.
. Correct detection of stop code for paper-tape search reverse.
. Correct word sign detection for TYPE19 and PUNCH19 commands.
. Correct drum line from which format codes are taken for TYPE19 and
PUNCH19 commands.
. Correct double sign-bit precession when formatting a carriage return
on output.
. Correct internal code for the "period" character; accept either "P"
or "." on input.
. Improve handling of I/O cancellation and initiation of an I/O while
another I/O is already in progress. Now the second I/O will cancel the
first one, but the OC register will be set to the "OR" of the two I/O
command codes.
. Use context-bound functions instead of dynamic drum object indexing
for output code precessors.
. Clear internal double-precision state variables when clearing the DP
registers (D=31 S=23).
. Apply additional manifest constants from Util module.
. Change FlipFlop.neonPersistence from 7 to 15 ms for lamp glow
calculations.
. Implement significant expansion of command tracing to the JavaScript
console. This required rearranging some code and introducing local
variables for reporting intermediate results, with a corresponding
negligible increase in overhead that is still present when tracing is
disabled. Eventually most of this tracing code can be removed.
. Disable command tracing by default -- can now be toggled dynamically
by double-clicking the version number on the Control Panel.