Improvements to paper-tape and typewriter I/O: v0.04a.
. Modify paper tape reader so that an I/O will simply hang if it is
initiated when the reader's input buffer is exhausted or no tape is in
the reader. Previously the reader would return a Stop code in this
condition.
. Correct typewriter keyboard handing so that keys for graphic
characters will echo to the paper and keys for control functions (e.g.,
Ctrl-C) will be passed through to the browser for default action.
. Correct this.busy vs this.ready tests in device drivers.
. Improve lamp brightness on Diagnostic Panel.
. Editorial changes to README.md and index.html.