v1.09: Miscellaneous Processor corrections and trace improvements.
. Correct timing problem with TEST for I/O READY not obeying the T
number in the command; insure that all immediate commands execute their
Transfer state for the correct number of word-times.
. Correct auto reload operation for SLOW IN I/O in Processor and Drum.
. Reinstate the ability for the TYPE AR command to terminate after
sending one word to the Typewriter even if the format string in line 03
does not have a STOP code (inadvertently deleted in v1.07b).
. Correct bug for commands D=31, S=31, CH=1&2 where L18 was being ADDED
to another line rather than being OR-ed to it (broken in v1.05b).
. Add current word-time to Processor trace output.
. Allow word-time counter and run-time lock to be reset in DiagPanel.
. Minor updates to web site files.